+++ product/warehouse/docker-compose.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-export CODEKVAST_VERSION=$(grep codekvastVersion ../../gradle.properties | tr -d [:space:] | cut -d= -f2)
-docker-compose $*